refactor(cli): extract tui/hook-log/logs into cli_tui.py (2b)
Recall-telemetry TUI + log viewers moved out of cli.py (only referenced
console/Config/json/time/Path). cli.py 2312 → 2100 LOC.
